@@ -15,7 +15,7 @@ from lbry.stream.descriptor import StreamDescriptor, sanitize_file_name
 
 class TestStreamDescriptor(AsyncioTestCase):
     async def asyncSetUp(self):
-        self.loop = asyncio.get_event_loop()
+        self.loop = asyncio.get_running_loop()
         self.key = b'deadbeef' * 4
         self.cleartext = os.urandom(20000000)
         self.tmp_dir = tempfile.mkdtemp()
